requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
scholarly>=1.7.11
tavily-python>=0.3.0
python-dotenv>=1.0.0
lxml>=4.9.0

